unisport/requirements.txt
Django==1.9.2
djangorestframework==3.3.2
Markdown==2.6.5
requests==2.9.1
wheel==0.24.0
